+++ .npmrc
@@ -1,1 +1,3 @@
 @digita-ai:registry=https://npm.pkg.github.com
+# add your github key and don't commit it
+# //npm.pkg.github.com/:_authToken=
