--- packages/service/.npmrc
@@ -1,1 +1,0 @@
-@digita-ai:registry=https://npm.pkg.github.com
