+++ packages/service/package.json
@@ -24,7 +24,9 @@
     "build:ts": "tsc",
     "build:watch": "tsc --watch",
     "clean": "rm -rf dist/",
-    "test": "jest",
+    "test:unit": "jest test/unit",
+    "test:integration": "jest test/integration --coverage=false",
+    "test": "npm run test:unit && npm run test:integration",
     "lint": "eslint . --ext .ts",
     "lint:fix": "eslint . --ext .ts --fix",
     "format": "prettier --write {src,test}/**/*.ts"
@@ -36,9 +38,9 @@
     "@digita-ai/handlersjs-core": "^0.11.0",
     "@digita-ai/handlersjs-http": "^0.11.0",
     "@inrupt/solid-client-authn-node": "^1.11.7",
-    "@janeirodigital/interop-authorization-agent": "^1.0.0-rc.6",
-    "@janeirodigital/interop-data-model": "^1.0.0-rc.6",
-    "@janeirodigital/interop-namespaces": "^1.0.0-rc.6",
+    "@janeirodigital/interop-authorization-agent": "^1.0.0-rc.8",
+    "@janeirodigital/interop-data-model": "^1.0.0-rc.8",
+    "@janeirodigital/interop-namespaces": "^1.0.0-rc.8",
     "@janeirodigital/sai-server-interfaces": "^0.1.0",
     "@janeirodigital/sai-server-mocks": "^0.1.0",
     "@janeirodigital/solid-redis-token-storage": "^0.0.5",
